requirements.txt
fastapi~=0.104.1
uvicorn[standard]
gunicorn
pydantic~=1.10.13
python-dotenv~=1.0.0
numpy~=1.23.0
slowapi~=0.1.9
requests~=2.31.0
joblib~=1.2.0
pandas~=1.5.3
xgboost~=2.1.0
scikit-learn~=1.0.2